refactor: reordered test in proper folder and proper file, with proper name
