--- packages/python/goatlib/tests/integration/routing/catchment/test_routing_workflow.py
@@ -1,191 +1,0 @@
-import logging
-import os
-from pathlib import Path
-
-import pytest
-from goatlib.analysis.network.network_processor import InMemoryNetworkProcessor
-from goatlib.routing.adapters.motis import create_motis_adapter
-from goatlib.routing.schemas.base import (
-    CatchmentAreaRoutingModePT,
-    Coordinates,
-)
-from goatlib.routing.schemas.catchment import (
-    CatchmentAreaType,
-    CatchmentRequest,
-)
-from goatlib.routing.schemas.catchment_area_transit import (
-    AccessEgressSettings,
-    TransitCatchmentAreaRequest,
-)
-
-logger = logging.getLogger(__name__)
-
-
-# fixture of a standard catchment request in Munich
-@pytest.fixture
-def munich_catchment_request() -> CatchmentRequest:
-    """Create a standard Munich catchment area request."""
-    return CatchmentRequest(
-        starting_points=[
-            Coordinates(lat=48.1351, lon=11.5820)  # Munich center
-        ],
-        cutoffs=[15, 30, 45],  # 15, 30, and 45 minute isochrones
-        type=CatchmentAreaType.point,
-    )
-
-
-@pytest.mark.asyncio
-@pytest.mark.network
-async def test_complete_motis_rust_workflow(munich_catchment_request: CatchmentRequest):
-    """Complete MOTIS + Rust workflow using the correct adapter interface."""
-
-    # Test data path
-    test_file = Path("/app/packages/python/goatlib/tests/data/network/network.parquet")
-    if not test_file.exists():
-        logger.info(f"❌ Test file not found: {test_file}")
-        pytest.skip(f"Test file not found: {test_file}")
-
-    # Step 1: Use MOTIS adapter directly with the interface
-    center = Coordinates(lat=48.1351, lon=11.5820)  # Munich center
-    motis_request = TransitCatchmentAreaRequest(
-        starting_points=[center],  # Munich
-        transit_modes=[
-            CatchmentAreaRoutingModePT.rail,
-            CatchmentAreaRoutingModePT.subway,
-            CatchmentAreaRoutingModePT.tram,
-        ],
-        cutoffs=[15],  # 15 minutes
-        access_settings=AccessEgressSettings.create_walk_settings(),
-        egress_settings=AccessEgressSettings.create_walk_settings(),
-    )
-
-    # Use the MOTIS client directly to get raw station data
-    adapter = create_motis_adapter()
-    try:
-        # Get raw MOTIS data using the converter functions
-        from goatlib.routing.adapters.motis.motis_converters import (
-            extract_bus_stations_for_buffering,
-            translate_to_motis_one_to_all_request,
-        )
-
-        # Convert our request to MOTIS format and call directly
-        motis_req = translate_to_motis_one_to_all_request(motis_request)
-        logger.info(f"MOTIS request: {motis_req}")
-
-        raw_motis_response = await adapter.motis_client.one_to_all(motis_req)
-        logger.info("MOTIS raw response received")
-
-        # Extract stations from the raw response
-        raw_stations = extract_bus_stations_for_buffering(raw_motis_response)
-        logger.info(f"Extracted {len(raw_stations)} transit stations")
-
-        if not raw_stations:
-            logger.info("❌ No stations found in MOTIS response")
-            pytest.skip("No station data from MOTIS")
-
-        # Show first few stations for debugging
-        for i, station in enumerate(raw_stations):
-            coords = station["coordinates"]  # [lon, lat]
-            logger.info(
-                f"   Station {i+1}: {station.get('name', 'Unknown')} at [{coords[1]:.4f}, {coords[0]:.4f}]"
-            )
-
-        # Convert to our format
-        stations_data = []
-        for station in raw_stations:
-            coords = station["coordinates"]  # [lon, lat]
-            stations_data.append(
-                {
-                    "name": station.get("name", "Unknown"),
-                    "lat": coords[1],  # latitude
-                    "lon": coords[0],  # longitude
-                    "transit_time": station.get("duration_minutes", 0),
-                }
-            )
-        # final_response = CatchmentResponse(last_mile_catchment=stations_data)
-
-    except Exception as e:
-        await adapter.motis_client.close()
-        logger.error(f"MOTIS API error: {e}")
-        pytest.skip(f"MOTIS API unavailable: {e}")
-
-    await adapter.motis_client.close()
-
-    # Step 2: Process with Rust routing using the fast functions
-    successful_routing = 0
-    total_reachable = 0
-
-    with InMemoryNetworkProcessor(input_path=str(test_file)) as proc:
-        # Load network around Munich - larger area to ensure we catch stations
-        subset_table = proc.load_network(
-            center=center, buffer_radius=15000.0
-        )  # 15km radius
-        logger.info(f"Loaded network subset: {subset_table}")
-
-        # Process all stations in batch using calculate_multiple_isochrones
-        station_coordinates = [
-            Coordinates(lat=station["lat"], lon=station["lon"])
-            for station in stations_data
-        ]
-
-        try:
-            # Create artificial nodes for all stations at once
-            result = proc.create_artificial_nodes_for_points(
-                station_coordinates, subset_table, search_radius_m=500.0
-            )
-
-            if isinstance(result, tuple):
-                output_path, artificial_node_ids = result
-                logger.info(
-                    f"Created: {len(artificial_node_ids)} artificial nodes for {len(station_coordinates)} stations"
-                )
-
-                # Use batch routing with calculate_multiple_isochrones
-                try:
-                    import fast_routing_py as routing
-
-                    network = routing.load_network(output_path)
-                    max_cost = 300  # 5 minutes in seconds
-
-                    # Use calculate_multiple_isochrones for all stations at once
-                    routing_results = network.calculate_multiple_isochrones(
-                        start_nodes=artificial_node_ids, max_cost=max_cost
-                    )
-
-                    # Process results
-                    for i, routing_result in enumerate(routing_results):
-                        if i < len(stations_data):
-                            station = stations_data[i]
-                            reachable = routing_result.reachable_nodes
-                            successful_routing += 1
-                            total_reachable += reachable
-
-                    # Cleanup
-                    if os.path.exists(output_path):
-                        os.unlink(output_path)
-
-                except Exception as e:
-                    logger.warning(f"❌ Batch routing failed: {e}")
-                    if os.path.exists(output_path):
-                        os.unlink(output_path)
-            else:
-                logger.warning("⚠️ Network processing failed")
-
-        except Exception as e:
-            logger.warning(f"❌ Station processing failed: {e}")
-
-    # Step 3: Results
-    success_rate = (
-        (successful_routing / len(stations_data) * 100) if stations_data else 0
-    )
-    logger.info(f"   MOTIS stations found: {len(stations_data)}")
-    logger.info(f"   Successful routing: {successful_routing}")
-    logger.info(f"   Success rate: {success_rate:.1f}%")
-    logger.info(f"   Total reachable nodes: {total_reachable:,}")
-
-    # Assertions
-    assert len(stations_data) > 0, "Should find transit stations from MOTIS"
-    assert (
-        successful_routing > 0
-    ), "Should successfully route from at least some stations"
-    assert total_reachable > 0, "Should find reachable nodes"
